# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"SkillWatch API","description":"\n## Observatoire du Marché de l'Emploi Data & IA\n\nSkillWatch agrège des données depuis 5 sources hétérogènes\npour répondre à : quelles compétences tech sont demandées,\npopulaires et bien rémunérées en France ?\n\n### Sources de données\n- **France Travail** : offres d'emploi en temps réel (API REST)\n- **Stack Overflow** : statistiques développeurs 2021-2025\n- **OpenClassrooms** : catalogue de formations tech (scraping)\n- **INSEE** : données démographiques par département (PostgreSQL)\n- **Apache Spark** : traitement Big Data des archives Stack Overflow\n\n### Authentification\nToutes les routes (sauf /health) nécessitent un token JWT.\nObtenez votre token via POST /auth/login.\n","contact":{"name":"SkillWatch","url":"https://github.com/NovaStarmax/skillwatch"},"version":"1.0.0"},"paths":{"/auth/login":{"post":{"tags":["auth"],"summary":"Obtenir un token JWT","description":"Authentification par username/password.\nRetourne un token Bearer valable 30 minutes.\n\nCredentials par défaut : **admin / skillwatch2025**\n\nUtilisation : `Authorization: Bearer {token}`","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/skills":{"get":{"tags":["skills"],"summary":"Liste tous les skills","description":"Retourne tous les skills connus avec leurs statistiques de marché agrégées (offres, usage développeurs, salaire moyen, formations). Triés par nombre d'offres décroissant, skills sans offres en fin de liste.","operationId":"list_skills_skills_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/SkillDetail"},"type":"array","title":"Response List Skills Skills Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/skills/{name}":{"get":{"tags":["skills"],"summary":"Détail d'un skill","description":"Retourne le détail complet d'un skill par son nom canonique. La recherche est insensible à la casse (ILIKE). Retourne 404 si le skill n'existe pas.","operationId":"get_skill_skills__name__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string","title":"Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SkillDetail"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/market/summary":{"get":{"tags":["market"],"summary":"Top 20 skills du marché","description":"Retourne les 20 skills les plus demandés en offres d'emploi France Travail. Chaque entrée inclut le salaire moyen (converti en EUR), la popularité Stack Overflow et le département avec le plus d'offres pour ce skill.","operationId":"market_summary_market_summary_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/MarketSummaryItem"},"type":"array","title":"Response Market Summary Market Summary Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/market/by-department":{"get":{"tags":["market"],"summary":"Offres par département","description":"Croise les offres France Travail avec les données démographiques INSEE. Retourne les 20 départements les plus actifs avec le ratio offres/million d'habitants pour neutraliser l'effet de la taille de population.","operationId":"by_department_market_by_department_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/DepartmentStats"},"type":"array","title":"Response By Department Market By Department Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/trainings":{"get":{"tags":["trainings"],"summary":"Liste toutes les formations","description":"Retourne toutes les formations OpenClassrooms scrapées, filtrées sur les domaines tech (Data, Développement, Systèmes & Réseaux, Cybersécurité). Triées par domaine puis titre.","operationId":"list_trainings_trainings_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Training"},"type":"array","title":"Response List Trainings Trainings Get"}}}}},"security":[{"HTTPBearer":[]}]}},"/trainings/skill/{skill_name}":{"get":{"tags":["trainings"],"summary":"Formations pour un skill","description":"Retourne les formations disponibles pour apprendre un skill spécifique. La recherche est insensible à la casse. Retourne une liste vide si aucune formation ne couvre ce skill.","operationId":"trainings_by_skill_trainings_skill__skill_name__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"skill_name","in":"path","required":true,"schema":{"type":"string","title":"Skill Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Training"},"title":"Response Trainings By Skill Trainings Skill  Skill Name  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["health"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"DepartmentStats":{"properties":{"dept_code":{"type":"string","title":"Dept Code"},"dept_name":{"type":"string","title":"Dept Name"},"population":{"type":"integer","title":"Population"},"job_count":{"type":"integer","title":"Job Count"},"jobs_per_million_hab":{"type":"number","title":"Jobs Per Million Hab"}},"type":"object","required":["dept_code","dept_name","population","job_count","jobs_per_million_hab"],"title":"DepartmentStats"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"LoginRequest"},"MarketSummaryItem":{"properties":{"skill":{"type":"string","title":"Skill"},"category":{"type":"string","title":"Category"},"job_offer_count":{"type":"integer","title":"Job Offer Count"},"developer_usage_count":{"type":"integer","title":"Developer Usage Count"},"avg_salary_eur":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Avg Salary Eur"},"training_count":{"type":"integer","title":"Training Count"},"top_dept":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Top Dept"},"top_dept_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Top Dept Name"},"top_dept_population":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Top Dept Population"}},"type":"object","required":["skill","category","job_offer_count","developer_usage_count","avg_salary_eur","training_count"],"title":"MarketSummaryItem"},"SkillDetail":{"properties":{"name":{"type":"string","title":"Name"},"category":{"type":"string","title":"Category"},"job_offer_count":{"type":"integer","title":"Job Offer Count"},"developer_usage_count":{"type":"integer","title":"Developer Usage Count"},"avg_salary_eur":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Avg Salary Eur"},"training_count":{"type":"integer","title":"Training Count"}},"type":"object","required":["name","category","job_offer_count","developer_usage_count","avg_salary_eur","training_count"],"title":"SkillDetail"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"TokenResponse"},"Training":{"properties":{"title":{"type":"string","title":"Title"},"domain":{"type":"string","title":"Domain"},"level":{"type":"string","title":"Level"},"duration_months":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Duration Months"},"provider":{"type":"string","title":"Provider"},"url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Url"}},"type":"object","required":["title","domain","level","duration_months","provider","url"],"title":"Training"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}